+++ jsonargparse_tests/deprecated_tests.py
@@ -119,18 +119,23 @@ class DeprecatedTests(unittest.TestCase):
         self.assertIsInstance(cfg_init['cal'], calendar.Calendar)
 
 
+class DeprecatedTempDirTests(TempDirTestCase):
+
     def test_parse_as_dict(self):
+        with open('config.json', 'w') as f:
+            f.write('{}')
         parser = ArgumentParser(parse_as_dict=True, default_meta=False)
         self.assertEqual({}, parser.parse_args([]))
         self.assertEqual({}, parser.parse_env([]))
         self.assertEqual({}, parser.parse_string('{}'))
         self.assertEqual({}, parser.parse_object({}))
-        with open('config.json', 'w') as f:
-            f.write('{}')
-        parser = ArgumentParser(parse_as_dict=True, default_meta=True)
+        self.assertEqual({}, parser.parse_path('config.json'))
+        self.assertEqual({}, parser.instantiate_classes({}))
+        self.assertEqual('{}\n', parser.dump({}))
+        parser.save({}, 'config.yaml')
+        with open('config.yaml') as f:
+            self.assertEqual('{}\n', f.read())
 
-
-class DeprecatedTempDirTests(TempDirTestCase):
 
     def test_ActionPath(self):
         os.mkdir(os.path.join(self.tmpdir, 'example'))